api/requirements.txt
flask
flask-cors
flask-socketio
flask_jwt_extended